Refactor Airport and Aircraft entities to use Long for ID fields; update related methods and repository

+++ src/main/java/com/keyin/airportapi/airport/Airport.java
@@ -1,13 +1,13 @@
 package com.keyin.airportapi.airport;
 
 import jakarta.persistence.*;
-import com.keyin.airportapi.city.*;
+import com.keyin.airportapi.city.City;
 
 @Entity
 @Table(name = "airport")
 public class Airport {
     @Id
-    private int airportId;
+    private Long airportId;
 
     private String airportName;
     private String areaCode;
@@ -19,7 +19,7 @@ public class Airport {
         // Empty constructor for testing. I read somewhere this is best practice.
     }
 
-    public Airport(int airportId, String airportName, String areaCode, City city) {
+    public Airport(Long airportId, String airportName, String areaCode, City city) {
         this.airportId = airportId;
         this.airportName = airportName;
         this.areaCode = areaCode;
@@ -27,7 +27,7 @@ public class Airport {
     }
 
     // Getters
-    public int getAirportId() {
+    public Long getAirportId() {
         return airportId;
     }
 
@@ -44,7 +44,7 @@ public class Airport {
     }
 
     // Setters
-    public void setAirportId(int airportId) {
+    public void setAirportId(Long airportId) {
         this.airportId = airportId;
     }
