+++ src/main/java/com/keyin/airportapi/airport/AirportRepository.java
@@ -2,6 +2,6 @@ package com.keyin.airportapi.airport;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AirportRepository extends CrudRepository<Airport, Integer> {
+public interface AirportRepository extends CrudRepository<Airport, Long> {
 
 }
