+++ src/main/java/com/keyin/airportapi/airport/AirportService.java
@@ -20,7 +20,7 @@ public class AirportService {
         return airports;
     }
 
-    public Optional<Airport> getAirportById(int id) {
+    public Optional<Airport> getAirportById(Long id) {
         return airportRepository.findById(id);
     }
 
@@ -28,7 +28,7 @@ public class AirportService {
         return airportRepository.save(airport);
     }
 
-    public Airport updateAirport(int id, Airport updatedAirport) {
+    public Airport updateAirport(Long id, Airport updatedAirport) {
         Optional<Airport> airportOptional = getAirportById(id);
         if (airportOptional.isPresent()) {
             Airport airport = airportOptional.get();
@@ -42,7 +42,7 @@ public class AirportService {
         }
     }
 
-    public void deleteAirport(int id) {
+    public void deleteAirport(Long id) {
         airportRepository.deleteById(id);
     }
 }
